+++ src/contexts/SettingsProvider.tsx
@@ -8,12 +8,15 @@ interface SidebarProviderProps {
 export const SettingsProvider: React.FC<SidebarProviderProps> = ({
     children,
 }) => {
+    const month = new Date().getMonth() + 1;
     const [openedSidebar, setOpenedSidebar] = useState<string>("text");
     const [hide, setHide] = useState<boolean>(false);
     const [hideAnnouncements, setHideAnnouncements] = useState<boolean>(true);
     const [showTutorial, setShowTutorial] = useState<boolean>(false);
     const [showSaveDialog, setShowSaveDialog] = useState<boolean>(false);
-    const [blankCanvas, setBlankCanvas] = useState<boolean>(false);
+    const [blankCanvas, setBlankCanvas] = useState<boolean>(
+        month === 10 ? true : false
+    );
     const [showExperimental, setShowExperimental] = useState<boolean>(false);
     const [openAll, setOpenAll] = useState<boolean>(false);
     const [openTextOption, setOpenTextOption] = useState<string>("name-tag");
@@ -27,7 +30,7 @@ export const SettingsProvider: React.FC<SidebarProviderProps> = ({
     const [loading, setLoading] = useState<number>(0);
 
     useEffect(() => {
-        const announcementCookie = localStorage.getItem("5.8.2-announcements");
+        const announcementCookie = localStorage.getItem("5.8.4-announcements");
         if (Number(announcementCookie) < 1) {
             setHideAnnouncements(false);
         }
@@ -51,8 +54,12 @@ export const SettingsProvider: React.FC<SidebarProviderProps> = ({
         if (!saveDialogCookie || saveDialogCookie === "true") {
             setShowSaveDialog(true);
         }
-        const blankCanvasCookie = localStorage.getItem("blankCanvas");
-        if (blankCanvasCookie === "true") {
+        const blankCanvasCookie = localStorage.getItem(
+            month === 10 ? "blankCanvasOctober" : "blankCanvas"
+        );
+        if (month === 10 && blankCanvasCookie === "false") {
+            setBlankCanvas(false);
+        } else if (blankCanvasCookie === "true") {
             setBlankCanvas(true);
         }
         const easySwitchEnabled = localStorage.getItem("easySwitchEnabled");